Refactor import path for ChallengeCard and ChallengeCardSkeleton components to improve project structure
