+++ src/components/groups/challenge/challenge-card/ChallengeCard.tsx
